Deal with circular deps
