+++ frontend/util.lua
@@ -1311,58 +1311,6 @@ function util.stringSearch(txt, str, case_sensitive, start_pos)
     return char_pos
 end
 
---[[--
-Natural sorting functions, for use with table.sort
-<http://notebook.kulchenko.com/algorithms/alphanumeric-natural-sorting-for-humans-in-lua>
---]]
--- Original implementation by Paul Kulchenko
---[[
-local function addLeadingZeroes(d)
-    local dec, n = string.match(d, "(%.?)0*(.+)")
-    return #dec > 0 and ("%.12f"):format(d) or ("%s%03d%s"):format(dec, #n, n)
-end
-function util.natsort(a, b)
-    return tostring(a):gsub("%.?%d+", addLeadingZeroes)..("%3d"):format(#b)
-           < tostring(b):gsub("%.?%d+", addLeadingZeroes)..("%3d"):format(#a)
-end
---]]
--- Hardened (but more expensive) implementation by Egor Skriptunoff, with an UTF-8 tweak by Paul Kulchenko
-local function natsort_conv(s)
-    local res, dot = "", ""
-    for n, m, c in tostring(s):gmatch("(0*(%d*))(.?)") do
-        if n == "" then
-            dot, c = "", dot..c
-        else
-            res = res..(dot == "" and ("%03d%s"):format(#m, m)
-                                   or "."..n)
-            dot, c = c:match("(%.?)(.*)")
-        end
-        res = res..c:gsub("[%z\1-\127\192-\255]", "\0%0")
-    end
-    return res
-end
--- The above conversion is *fairly* expensive,
--- and table.sort ensures that it'll be called on identical strings multiple times,
--- so keeping a cache of massaged strings makes sense.
--- <https://github.com/koreader/koreader/pull/10023#discussion_r1069776657>
--- Rely on LRU to avoid explicit cache maintenance concerns (at the cost of a bit of memory).
--- The extra persistence this affords us also happens to help with the FM use-case ;).
-local lru = require("ffi/lru")
-local natsort_cache = lru.new(G_defaults:readSetting("DNATURAL_SORT_CACHE_SIZE"), nil, false)
-function util.natsort(a, b)
-    local ca, cb = natsort_cache:get(a), natsort_cache:get(b)
-    if not ca then
-        ca = natsort_conv(a)
-        natsort_cache:set(a, ca)
-    end
-    if not cb then
-        cb = natsort_conv(b)
-        natsort_cache:set(b, cb)
-    end
-
-    return ca < cb or ca == cb and a < b
-end
-
 local WrappedFunction_mt = {
     __call = function(self, ...)
         if self.before_callback then
